The first implementation
Features
1. Monitor a user-selected folder
2. Replace given strings separated with semicomma with a given string.
3. Log history.

This program is designed to modified NC programs for CNC machines. Because different CNCs may assign various M codes for the same work, i.e. M130 and M199, it is better to replace the M codes with Macro variable, such as M#222. Users can assign 130 and 199 to #222 in difference machines.